Move the ParsingService to the parser package, because that just seems a sensible thing to do.
