+++ src/main/java/com/gnopai/ji65/parser/statement/IncludeSourceFileDirectiveParser.java
@@ -1,7 +1,7 @@
 package com.gnopai.ji65.parser.statement;
 
-import com.gnopai.ji65.ParsingService;
 import com.gnopai.ji65.parser.Parser;
+import com.gnopai.ji65.parser.ParsingService;
 import com.gnopai.ji65.scanner.Token;
 
 import javax.inject.Inject;
