+++ src/main/java/com/gnopai/ji65/parser/statement/MacroDirectiveParser.java
@@ -1,8 +1,8 @@
 package com.gnopai.ji65.parser.statement;
 
-import com.gnopai.ji65.ParsingService;
 import com.gnopai.ji65.parser.Macro;
 import com.gnopai.ji65.parser.Parser;
+import com.gnopai.ji65.parser.ParsingService;
 import com.gnopai.ji65.scanner.Token;
 import com.gnopai.ji65.scanner.TokenType;
