+++ src/test/java/com/gnopai/ji65/parser/ParserTestUtil.java
@@ -1,6 +1,5 @@
 package com.gnopai.ji65.parser;
 
-import com.gnopai.ji65.ParsingService;
 import com.gnopai.ji65.parser.statement.Statement;
 import com.gnopai.ji65.scanner.Token;
 import com.gnopai.ji65.scanner.TokenType;
