+++ src/test/java/com/gnopai/ji65/parser/statement/IncludeSourceFileDirectiveParserTest.java
@@ -1,7 +1,7 @@
 package com.gnopai.ji65.parser.statement;
 
 import com.gnopai.ji65.DirectiveType;
-import com.gnopai.ji65.ParsingService;
+import com.gnopai.ji65.parser.ParsingService;
 import com.gnopai.ji65.util.ErrorHandler;
 import org.junit.jupiter.api.Test;
