+++ src/test/java/com/gnopai/ji65/parser/statement/MacroCallParseletTest.java
@@ -1,9 +1,9 @@
 package com.gnopai.ji65.parser.statement;
 
 import com.gnopai.ji65.InstructionType;
-import com.gnopai.ji65.ParsingService;
 import com.gnopai.ji65.parser.Macro;
 import com.gnopai.ji65.parser.ParseException;
+import com.gnopai.ji65.parser.ParsingService;
 import com.gnopai.ji65.scanner.Token;
 import com.gnopai.ji65.util.ErrorHandler;
 import lombok.Builder;
